+++ src/features/auth/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useExchangeAuthToken } from "@/features/auth/hooks/useExchangeAuthToken";
+export { useRegistrationStatus } from "@/features/auth/hooks/useRegistrationStatus";
