+++ src/features/auth/index.ts
@@ -1,3 +1,4 @@
 export * from "@/features/auth/hooks";
 export * from "@/features/auth/queries";
+export * from "@/features/auth/types";
 export { getGoogleLoginUrl } from "@/features/auth/utils";
