+++ src/features/auth/types/index.ts
@@ -1,0 +1,12 @@
+export type AuthTokenRequest = {
+  code: string;
+};
+
+export type AuthTokenResponse = {
+  accessToken: string;
+  isChecklistRegistered: boolean;
+  isOnboarded: boolean;
+  isRoommatePreferenceRegistered: boolean;
+  userId: number;
+  username: string;
+};
